Allow objects with a url property.
Instead of only parsing strings of urls, it can now parse an object
that has a url property. This allows the user to pass properties along
with the url object that can then later be used when the data is
returned. Useful if you need to do something different for each url
that has been scraped. I also wrote tests and corrected another test
that was for some reason always passed, although it ought to fail.